Refactor traffic_control views into a package
All traffic_control views were in one module. Split the view classes
into a package that is split into modules the same way as models are.

There's no actual changes other moving the view classes around and
fixing the imports in urls.

Refs LIIK-112
